Reduce the complexity of subscribe to an websocket topic
Before, we have to initilize a GenServer instance, use WebSockex class
to subscribe a topic, and use Phoenix PubSub to broadcast the message to
listeners.

Now, we make a boilerplate to handle these steps, no need more complex
steps as before. Just define a list of topic would you like to
subscribe, and a method to handle those topic messages